Move `useModal` state to a provider
